// ep-frontend/src/api-client/models/MailDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MailDTO
 */
export interface MailDTO {
    /**
     * 
     * @type {string}
     * @memberof MailDTO
     */
    from: string;
    /**
     * 
     * @type {string}
     * @memberof MailDTO
     */
    subject: string;
    /**
     * 
     * @type {string}
     * @memberof MailDTO
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof MailDTO
     */
    text: string;
}

/**
 * Check if a given object implements the MailDTO interface.
 */
export function instanceOfMailDTO(value: object): boolean {
    if (!('from' in value)) return false;
    if (!('subject' in value)) return false;
    if (!('name' in value)) return false;
    if (!('text' in value)) return false;
    return true;
}

export function MailDTOFromJSON(json: any): MailDTO {
    return MailDTOFromJSONTyped(json, false);
}

export function MailDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): MailDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'from': json['from'],
        'subject': json['subject'],
        'name': json['name'],
        'text': json['text'],
    };
}

export function MailDTOToJSON(value?: MailDTO | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'from': value['from'],
        'subject': value['subject'],
        'name': value['name'],
        'text': value['text'],
    };
}